Move profile controller and tab to a feature directory
